# api/openapi.yaml
openapi: 3.0.0
info:
  title: Zymurgauge API
  version: 0.1.0
servers:
  - url: http://{host}:{port}/{basePath}
    description: The production API server
    variables:
      host:
        enum:
        - 'localhost'
        - 'zymurgauge.local'
        default: localhost
      port:
        enum:
          - '8080'
        default: '8080'
      basePath:
        enum:
        - 'v1'
        default: v1
paths:
  '/chambers':
    get:
      description: Returns all chambers
      operationId: getChambers
      responses:
        '200':
          description: OK response with list of chambers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chamber'
              examples:
                chambers:
                  $ref: '#/components/examples/chambers'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
    post:
      description: Saves a chamber
      operationId: saveChamber
      requestBody:
        description: Chamber to save
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Chamber'
            examples:
              newChamber:
                $ref: '#/components/examples/newChamber'
              updateChamber:
                $ref: '#/components/examples/updateChamber'
      responses:
        '200':
          description: OK response with saved chamber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chamber'
              examples:
                chamber:
                  $ref: '#/components/examples/chamber'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
  '/chambers/{id}':
    get:
      description: Returns a single chamber by id
      operationId: getChamberByID
      parameters:
        - name: id
          in: path
          description: ID of chamber to get
          required: true
          schema:
            type: string
            format: uuid
          example: 96f58a65-03c0-49f3-83ca-ab751bbf3768
      responses:
        '200':
          description: OK response with a chamber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chamber'
              examples:
                chamber:
                  $ref: '#/components/examples/chamber'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/chamberNotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
    delete:
      description: Deletes a single chamber by id
      operationId: deleteChamberByID
      parameters:
        - name: id
          in: path
          description: ID of chamber to delete
          required: true
          schema:
            type: string
            format: uuid
          example: 96f58a65-03c0-49f3-83ca-ab751bbf3768
      responses:
        '200':
          description: OK response
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/chamberNotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
  '/chambers/{id}/start':
    post:
      description: Starts a fermentation at the given step
      operationId: startFermentation
      parameters:
        - name: id
          in: path
          description: ID of the chamber to start fermentation in
          required: true
          schema:
            type: string
            format: uuid
          example: 96f58a65-03c0-49f3-83ca-ab751bbf3768
        - name: step
          in: query
          description: step of fermentation
          required: true
          schema:
            type: number
            format: integer
          example: 1
      responses:
        '200':
          description: OK response with a chamber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
              examples:
                chamber:
                  $ref: '#/components/examples/success'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/chamberNotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
  '/chambers/{id}/stop':
    post:
      description: Stops a fermentation
      operationId: stopFermentation
      parameters:
        - name: id
          in: path
          description: ID of the chamber to stop fermentation in
          required: true
          schema:
            type: string
            format: uuid
          example: 96f58a65-03c0-49f3-83ca-ab751bbf3768
      responses:
        '200':
          description: OK response with a chamber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
              examples:
                chamber:
                  $ref: '#/components/examples/success'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/chamberNotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
  '/thermometers':
    get:
      description: Returns all thermometer ids
      operationId: getThermometers
      responses:
        '200':
          description: OK response with list of thermometer ids
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["28-0000071cbc72","28-0000041ab222"]
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
  '/batches':
    get:
      description: Returns all batches
      operationId: getBatches
      responses:
        '200':
          description: OK response with list of batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Batch'
              examples:
                error:
                  $ref: '#/components/examples/batches'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
  '/batches/{id}':
    get:
      description: Returns a single batch by id
      operationId: getBatchByID
      parameters:
        - name: id
          in: path
          description: ID of batch to get
          required: true
          schema:
            type: string
            format: uuid
          example: KBTM3F9soO5TtbAx0A5mBZTAUsNZyg
      responses:
        '200':
          description: OK response with a batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
              examples:
                error:
                  $ref: '#/components/examples/success'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/batchNotFoundError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                error:
                  $ref: '#/components/examples/internalServerError'
components:
  schemas:
    Chamber:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        thermometerId:
          type: string
        chillerPin:
          type: string
        heaterPin:
          type: string
        chillerKp:
          type: number
          format: double
        chillerKi:
          type: number
          format: double
        chillerKd:
          type: number
          format: double
        heaterKp:
          type: number
          format: double
        heaterKi:
          type: number
          format: double
        heaterKd:
          type: number
          format: double
        modTime:
          type: string
          format: date-time
    Batch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        modTime:
          type: string
          format: date-time
    Status:
      type: object
      properties:
        message:
          type: string
    Error:
      type: object
      properties:
        errMessage:
          type: string
  examples:
    chamber:
      value:
        id: 96f58a65-03c0-49f3-83ca-ab751bbf3768
        name: My Fermentation Chamber
        thermometerId: 28-0000071cbc72
        chillerPin: "1"
        heaterPin: "2"
        chillerKp: 1.0
        chillerKi: 2.0
        chillerKd: 3.0
        heaterKp: 1.2
        heaterKi: 2.3
        heaterKd: 3.4
        currentBatch: 
          id: KBTM3F9soO5TtbAx0A5mBZTAUsNZyg
          name: My Pale Ale
          fermentation: 
              Name: Imported
              Steps:
                - type: "Primary"
                  actualTime: 1635573600000
                  stepTemp: 19.4
                  stepTime: 4
                - type: Secondary
                  actualTime: 1635919200000
                  stepTime: 10
                  stepTemp: 19.
                - type": "Conditioning"
                  actualTime: 1636786800000
                  stepTime: 30
                  stepTemp: 18.3
        currentFermentationStep: 2
        modTime: 2021-10-28 09:54:07.155132 -0600 MDT m=+0.000777430
    chambers:
      value:
        - id: 96f58a65-03c0-49f3-83ca-ab751bbf3768
          name: My Fermentation Chamber
          thermometerId: 28-0000071cbc72
          chillerPin: "1"
          heaterPin: "2"
          chillerKp: 1.0
          chillerKi: 2.0
          chillerKd: 3.0
          heaterKp: 1.2
          heaterKi: 2.3
          heaterKd: 3.4
          currentBatch: 
            id: KBTM3F9soO5TtbAx0A5mBZTAUsNZyg
            name: My Pale Ale
            fermentation: 
                Name: Imported
                Steps:
                  - type: "Primary"
                    actualTime: 1635573600000
                    stepTemp: 19.4
                    stepTime: 4
                  - type: Secondary
                    actualTime: 1635919200000
                    stepTime: 10
                    stepTemp: 19.
                  - type": "Conditioning"
                    actualTime: 1636786800000
                    stepTime: 30
                    stepTemp: 18.3
          currentFermentationStep: 2
          modTime: 2021-10-28 09:54:07.155132 -0600 MDT m=+0.000777430
        - id: dd2610fe-95fc-45f3-8dd8-3051fb1bd4c1
          name: My Fermentation Chamber
          thermometerId: 28-0000041ab222
          chillerPin: 1
          heaterPin: 2
          chillerKp: 1.0
          chillerKi: 2.0
          chillerKd: 3.0
          heaterKp: 1.2
          heaterKi: 2.3
          heaterKd: 3.4
          modTime: 2021-10-28 09:54:07.155132 -0600 MDT m=+0.000777430
    newChamber:
      value:
        name: My Fermentation Chamber
        thermometerId: 28-0000071cbc72
        chillerPin: "1"
        heaterPin: "2"
        chillerKp: 1.0
        chillerKi: 2.0
        chillerKd: 3.0
        heaterKp: 1.2
        heaterKi: 2.3
        heaterKd: 3.4
    updateChamber:
      value:
        id: 96f58a65-03c0-49f3-83ca-ab751bbf3768
        name: My Fermentation Chamber
        thermometerId: 28-0000071cbc72
        chillerPin: "1"
        heaterPin: "2"
        chillerKp: 1.0
        chillerKi: 2.0
        chillerKd: 3.0
        heaterKp: 1.2
        heaterKi: 2.3
        heaterKd: 3.4
    batch:
      value:
        id: KBTM3F9soO5TtbAx0A5mBZTAUsNZyg
        name: My Pale Ale
        fermentation: 
            Name: Imported
            Steps:
              - type: "Primary"
                actualTime: 1635573600000
                stepTemp: 19.4
                stepTime: 4
              - type: Secondary
                actualTime: 1635919200000
                stepTime: 10
                stepTemp: 19.
              - type": "Conditioning"
                actualTime: 1636786800000
                stepTime: 30
                stepTemp: 18.3
    batches:
      value:
        - id: KBTM3F9soO5TtbAx0A5mBZTAUsNZyg
          name: My Pale Ale
          fermentation: 
              Name: Imported
              Steps:
                - type: "Primary"
                  actualTime: 1635573600000
                  stepTemp: 19.4
                  stepTime: 4
                - type: Secondary
                  actualTime: 1635919200000
                  stepTime: 10
                  stepTemp: 19.
                - type: "Conditioning"
                  actualTime: 1636786800000
                  stepTime: 30
                  stepTemp: 18.3
        - id: qbFkFfeaJdL1ZibNLTu2lfRjIii1qW
          name: Stout
          fermentation: 
              Name: Imported
              Steps:
                - type: "Primary"
                  actualTime: 1635573600000
                  stepTemp: 19.4
                  stepTime: 4
                - type": "Conditioning"
                  actualTime: 1636786800000
                  stepTime: 30
                  stepTemp: 18.3
    success:
      value:
        message: Success
    chamberNotFoundError:
      value:
        error: chamber '96f58a65-03c0-49f3-83ca-ab751bbf3768' not found
    batchNotFoundError:
      value:
        error: batch 'KBTM3F9soO5TtbAx0A5mBZTAUsNZyg' not found
    internalServerError:
      value:
        error: Internal Server Error